Differential Equations

Euler Method

Qaylla.euler — Function

euler(F::Function,
      t₀::Number,
      y₀::Number,
      tₙ::Number,
      nh::Number,
      return_values::Bool=false)

Computes the Euler aproximation for y(t) in tₙ.

Arguments

F(t,y)::Function: The y'(t).

t₀::Number: The initial point.

tₙ::Number: The final point.

nh::Number: If 1≤nh, nh is the number of subintervals beetwen t₀ and tₙ,

else nh is the length of the subintervals beetwen t₀ and tₙ.

return_values::Bool: If is true return tvalues=[t₀,t₁...,tₙ], yvalues=[y₀,y₁...,yₙ], else return yₙ.

Example

julia> using Qaylla

julia> F(t,y)=t-y
F (generic function with 1 method)

julia> euler(F,0,2,1,5)
0.9830400000000001

julia> euler(F,0,2,1,5,true)
(Any[0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], Any[2, 1.6, 1.32, 1.1360000000000001, 1.0288000000000002, 0.9830400000000001])

julia> euler(F,0,2,1,0.2,true)
(Any[0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], Any[2, 1.6, 1.32, 1.1360000000000001, 1.0288000000000002, 0.9830400000000001])