Integrales

trapeze method

Qaylla.trapeze — Function

trapeze(f::Function,a::Number,b::Number,n::Int64=1)::Float64

Computes the integral of f applying the composite trapeze method.

Example

julia> using Qaylla

julia> f(x)=sin(x)
f (generic function with 1 method)

julia> a=0
0

julia> b=pi/2
1.5707963267948966

julia> n=4
4

julia> trapeze(f,a,b,n)
0.9871158009727754

Simpson method

Qaylla.simpson — Function

simpson(f::Function,a::Number,b::Number,n::Int64=2)::Float64

Computes the integral of f applying the composite Simpson method.

Example

julia> using Qaylla

julia> f(x)=sin(x)
f (generic function with 1 method)

julia> a=0
0

julia> b=pi/2
1.5707963267948966

julia> n=4
4

julia> simpson(f,a,b,n)
1.0001345849741938

Romberg method

Qaylla.romberg — Function

romberg(f::Function,a::Number,b::Number,n::Int64=4)::Float64

Computes the integral of f applying the Romberg method.

Example

julia> using Qaylla

julia> f(x)=sin(x)
f (generic function with 1 method)

julia> a=0
0

julia> b=pi/2
1.5707963267948966

julia> n=4
4

julia> romberg(f,a,b,n)
1.0000000081440203

Newton-Cotes methods

Qaylla.newton_cotes — Function

newton_cotes(f::Function,a::Number,b::Number,n::Int64,closed::Bool=true)::Float64

Computes the integral of f applying the Newton-Cotes methods. If closed=true 1≤n≤4, else 0≤n≤3.

Example

julia> using Qaylla

julia> f(x)=sin(x)
f (generic function with 1 method)

julia> a=0
0

julia> b=pi/2
1.5707963267948966

julia> newton_cotes(f,a,b,1)
0.7853981633974483

julia> newton_cotes(f,a,b,2)
1.0022798774922104

julia> newton_cotes(f,a,b,3)
1.001004923314279

julia> newton_cotes(f,a,b,4)
0.9999915654729927

julia> newton_cotes(f,a,b,0,false)
1.1107207345395915

julia> newton_cotes(f,a,b,1,false)
1.0728738432865557

julia> newton_cotes(f,a,b,2,false)
0.9979892924561773

julia> newton_cotes(f,a,b,3,false)
0.9986082958707345

trapeze2d method

Qaylla.trapeze2d — Method

trapeze2d(F::Function,
          a::Number,
          b::Number,
          c::Number,
          d::Number,
          n::Int,
          m::Int)::Float64

Computes the integral of F in a rectangular region the plain XY, applying the composite trapeze method.

Arguments

F(x,y)::Function: F(x,y) in the function to integrate.

a::Number: The lower limit of the region in x.

b::Number: The upper limit of the region in x.

c::Number: The lower limit of the region in y.

d::Number: The upper limit of the region in y.

n::Number: The number of subintervals beetwen a and b.

m::Number: The number of subintervals beetwen c and d.

Example

julia> using Qaylla

julia> F(x,y)=x^2 + sin(y)
F (generic function with 1 method)

julia> trapeze2d(F,0,1,0,1,2,2)
0.8250805155040757

julia> trapeze2d(F,0,1,0,1,4,2)
0.7938305155040756

julia> trapeze2d(F,0,1,0,1,2,4)
0.8323009375715021

julia> trapeze2d(F,0,1,0,1,4,4)
0.8010509375715024

Qaylla.trapeze2d — Method

trapeze2d(F::Function,
          a::Number,
          b::Number,
          c::Function,
          d::Function,
          n::Int,
          m::Int)::Float64

Computes the integral of F in a not rectangular region the plain XY, applying the composite trapeze method.

Arguments

F(x,y)::Function: F(x,y) in the function to integrate.

a::Number: The lower limit of the region in x.

b::Number: The upper limit of the region in x.

c(x)::Function: The lower limit of the region in y.

d(x)::Function: The upper limit of the region in y.

n::Number: The number of subintervals beetwen a and b.

m::Number: The number of subintervals beetwen c(x) and d(x).

Example

julia> using Qaylla

julia> F(x,y)=x^2 + sin(y)
F (generic function with 1 method)

julia> c(x)=x
c (generic function with 1 method)

julia> d(x)=2*x
d (generic function with 1 method)

julia> trapeze2d(F,0,1,c,d,2,2)
0.7143707767687785

julia> trapeze2d(F,0,1,c,d,4,2)
0.6526600971423198

julia> trapeze2d(F,0,1,c,d,2,4)
0.7187860972222649

julia> trapeze2d(F,0,1,c,d,4,4)
0.6563466180940812

Simpson2d method

Qaylla.simpson2d — Method

simpson2d(F::Function,
          a::Number,
          b::Number,
          c::Number,
          d::Number,
          n::Int,
          m::Int)::Float64

Computes the integral of F in a rectangular region the plain XY, applying the composite Simpson method.

Arguments

F(x,y)::Function: F(x,y) in the function to integrate.

a::Number: The lower limit of the region in x.

b::Number: The upper limit of the region in x.

c::Number: The lower limit of the region in y.

d::Number: The upper limit of the region in y.

n::Number: The number of subintervals beetwen a and b.

m::Number: The number of subintervals beetwen c and d.

Example

julia> using Qaylla

julia> F(x,y)=x^2 + sin(y)
F (generic function with 1 method)

julia> simpson2d(F,0,1,0,1,2,2)
0.793195523204118

julia> simpson2d(F,0,1,0,1,2,4)
0.7930410782606442

julia> simpson2d(F,0,1,0,1,4,2)
0.793195523204118

julia> simpson2d(F,0,1,0,1,4,4)
0.7930410782606443

julia> simpson2d(F,0,1,0,1,100,100)
0.7930310274907315

Qaylla.simpson2d — Method

simpson2d(F::Function,
          a::Number,
          b::Number,
          c::Function,
          d::Function,
          n::Int,
          m::Int)::Float64

Computes the integral of F in a not rectangular region the plain XY, applying the composite Simpson method.

Arguments

F(x,y)::Function: F(x,y) in the function to integrate.

a::Number: The lower limit of the region in x.

b::Number: The upper limit of the region in x.

c(x)::Function: The lower limit of the region in y.

d(x)::Function: The upper limit of the region in y.

n::Number: The number of subintervals beetwen a and b.

m::Number: The number of subintervals beetwen c(x) and d(x).

Example

julia> using Qaylla

julia> F(x,y)=x^2 + sin(y)
F (generic function with 1 method)

julia> c(x)=x
c (generic function with 1 method)

julia> d(x)=2*x
d (generic function with 1 method)

julia> simpson2d(F,0,1,c,d,2,2)
0.6343236523627963

julia> simpson2d(F,0,1,c,d,4,2)
0.6367323875670757

julia> simpson2d(F,0,1,c,d,2,4)
0.6342654852512393

julia> simpson2d(F,0,1,c,d,4,4)
0.6366813209795263