Differential Equations

Euler Method

Qaylla.euler — Method

euler(F::Function,
      a::Number,
      y_0::Number,
      b::Number,
      nh::Number)::Tuple{Vector{Float64}, Vector{Float64}}

Computes the Euler aproximation for y(t), when a≤t≤b.

Arguments

F(t,y)::Function: The y'(t).

a::Number: The initial point.

y_0::Number: The y value in a.

b::Number: The final point.

nh::Number: If 1≤nh, nh is the number of subintervals beetwen a and b,

else nh is the length of the subintervals beetwen a and b.

Return

t_values::Array{Float64}: [t₀,t₁...,tₙ], where tᵢ=a + ih

w_values::Array{Float64}: [w₀,w₁...,wₙ], where wᵢ≈yᵢ=y(tᵢ)

Example

julia> using Qaylla

julia> F(t,y)=t-y
F (generic function with 1 method)

julia> euler(F,0,2,1,5)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.6, 1.32, 1.1360000000000001, 1.0288000000000002, 0.9830400000000001])

julia> euler(F,0,2,1,0.2)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.6, 1.32, 1.1360000000000001, 1.0288000000000002, 0.9830400000000001])

Modified Euler Method

Qaylla.modified_euler — Method

modified_euler(F::Function,
               a::Number,
               y_0::Number,
               b::Number,
               nh::Number)::Tuple{Vector{Float64}, Vector{Float64}}

Computes the modified Euler aproximation for y(t), when a≤t≤b.

Arguments

F(t,y)::Function: The y'(t).

a::Number: The initial point.

y_0::Number: The y value in a.

b::Number: The final point.

nh::Number: If 1≤nh, nh is the number of subintervals beetwen a and b,

else nh is the length of the subintervals beetwen a and b.

Return

t_values::Array{Float64}: [t₀,t₁...,tₙ], where tᵢ=a + ih

w_values::Array{Float64}: [w₀,w₁...,wₙ], where wᵢ≈yᵢ=y(tᵢ)

Example

julia> using Qaylla

julia> F(t,y)=t-y
F (generic function with 1 method)

julia> modified_euler(F,0,2,1,5)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.66, 1.4171999999999998, 1.2541039999999999, 1.15636528, 1.1122195296])

julia> modified_euler(F,0,2,1,0.2)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.66, 1.4171999999999998, 1.2541039999999999, 1.15636528, 1.1122195296])

Runge Kutta of Order 2 Method

Qaylla.runge_kutta_order_2 — Method

runge_kutta_order_2(F::Function,
                    a::Number,
                    y_0::Number,
                    b::Number,
                    nh::Number)::Tuple{Vector{Float64}, Vector{Float64}}

Computes the Runge Kutta approximation of order 2 for y(t), when a≤t≤b.

Arguments

F(t,y)::Function: The y'(t).

a::Number: The initial point.

y_0::Number: The y value in a.

b::Number: The final point.

nh::Number: If 1≤nh, nh is the number of subintervals beetwen a and b,

else nh is the length of the subintervals beetwen a and b.

Return

t_values::Array{Float64}: [t₀,t₁...,tₙ], where tᵢ=a + ih

w_values::Array{Float64}: [w₀,w₁...,wₙ], where wᵢ≈yᵢ=y(tᵢ)

Example

julia> using Qaylla

julia> F(t,y)=t-y
F (generic function with 1 method)

julia> runge_kutta_order_2(F,0,2,1,5)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.66, 1.4171999999999998, 1.2541039999999999, 1.15636528, 1.1122195296])

julia> runge_kutta_order_2(F,0,2,1,0.2)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.66, 1.4171999999999998, 1.2541039999999999, 1.15636528, 1.1122195296])

Heun Method

Qaylla.heun — Method

heun(F::Function,
     a::Number,
     y_0::Number,
     b::Number,
     nh::Number)::Tuple{Vector{Float64}, Vector{Float64}}

Computes the Heun aproximation for y(t), when a≤t≤b.

Arguments

F(t,y)::Function: The y'(t).

a::Number: The initial point.

y_0::Number: The y value in a.

b::Number: The final point.

nh::Number: If 1≤nh, nh is the number of subintervals beetwen a and b,

else nh is the length of the subintervals beetwen a and b.

Return

t_values::Array{Float64}: [t₀,t₁...,tₙ], where tᵢ=a + ih

w_values::Array{Float64}: [w₀,w₁...,wₙ], where wᵢ≈yᵢ=y(tᵢ)

Example

julia> using Qaylla

julia> F(t,y)=t-y
F (generic function with 1 method)

julia> heun(F,0,2,1,5)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.656, 1.410645333333333, 1.2460483128888886, 1.1475648854850369, 1.1032064529170835])

julia> heun(F,0,2,1,0.2)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.656, 1.410645333333333, 1.2460483128888886, 1.1475648854850369, 1.1032064529170835])

Runge Kutta of Order 4 Method

Qaylla.runge_kutta_order_4 — Method

runge_kutta_order_4(F::Function,
                    a::Number,
                    y_0::Number,
                    b::Number,
                    nh::Number)::Tuple{Vector{Float64}, Vector{Float64}}

Computes the Runge Kutta approximation of order 4 for y(t), when a≤t≤b.

Arguments

F(t,y)::Function: The y'(t).

a::Number: The initial point.

y_0::Number: The y value in a.

b::Number: The final point.

nh::Number: If 1≤nh, nh is the number of subintervals beetwen a and b,

else nh is the length of the subintervals beetwen a and b.

Return

t_values::Array{Float64}: [t₀,t₁...,tₙ], where tᵢ=a + ih

w_values::Array{Float64}: [w₀,w₁...,wₙ], where wᵢ≈yᵢ=y(tᵢ)

Example

julia> using Qaylla

julia> F(t,y)=t-y
F (generic function with 1 method)

julia> runge_kutta_order_4(F,0,2,1,5)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.6562000000000001, 1.4109728133333335, 1.2464504747031113, 1.1480038853219274, 1.103655714375906])

julia> runge_kutta_order_4(F,0,2,1,0.2)
([0.0, 0.2, 0.4, 0.6000000000000001, 0.8, 1.0], [2.0, 1.6562000000000001, 1.4109728133333335, 1.2464504747031113, 1.1480038853219274, 1.103655714375906])